🎨 lazy app creation & store connections on app
app.js is now lazy in that it exports a function that
creates an express app on execution. This allows better
control of side effects for integration testing.

Both client connections required by the app (postmark
and postgres) are instantiated on app creation and stored
on the root Application object instead of as local
variables in various files. These client connection objects
are then passed as arguments to the code that requires
them instead of required. This makes testing simpler.

Related to #190