🔨 Replace direct API calls on client side
Use the custom HTTP API library to make API calls from
the React app instead of directly using `fetch` or the
`adminPost`/`adminGet` wrappers.

In some cases this simplifies the code. These changes are
part of the effort to migrate towards the new API design.
They are therefore incomplete, but functioning.

Related to #264