🐛 fix weird activities page bug
Turns out to be a styling bug. I don't understand
the mechanics of it, but I can no longer reproduce
the bug.

It was probably that the absolute positioning on the
input elements was somehow messing with the flexbox
layouts. I removed the absolute positioning, since it
looked to be having no effect anyway (except for
causing this bug).

Also seems to be unique to Chrome, as was unable
to reproduce on Firefox.

Related to #353