🐛 use acceptable PNG filtering algorithm on default png
For some reason this PNG was encoded using a non-standard
filtering algorithm, causing `pdfmake` and its underlying
dependency to throw an error.

I opened the PNG in Inkscape and re-exported it with the
default export settings, seems to work now.

See here for more: https://github.com/bpampuch/pdfmake/issues/879