🗃 update the initial migrations
Split up the initial migration into several files for readability.

Rename the seed files with a ".seed.sql" extension, and make them more
consistent.

Introduce supertype/subtype hierarchy where "organisations" are the
supertype and "community businesses" and "funding bodies" are the
subtype.

This is because they require different information to be stored, and
interact with different entities in different ways. For example,
entities related to visits and volunteers are concerned only with
community businesses, not funding bodies.

Any funding bodies that also want access to the "community business"
oriented features will need to be offered some mechanism to populate the
community_business table with a row referencing their organisation.

Related to #1