🐛 remove homepage to stop CRA prefixing generated asset URLs
For some reason (🙄) CRA uses the value of `homepage` in the package.json
as a prefix on the static asset URLs (e.g. JS scripts) in the HTML file
that it generates.

If this doesn't actually correspond to the directory that the files are
served from, your assets aren't served.