🐛 fix bug by tracking sorted column by index instead of title (#129)
Fixes #128 

### Changes
* `sortBy` state fragment tracks column index (`number`) instead of column name (`string`) since the latter is not fixed but the former is
* Use default empty initial state for `tableData` so components don't have to deal with the null case
* Render loader while data is being fetched to avoid rendering `NO DATA AVAILABLE` view unnecessarily

Note:
I think the hooks should be refactored to allow this all to happen a little more easily and with quite so much boilerplate. We can certainly build some higher-level hooks on top of the ones we already have. However I decided not to do this in this PR in the interests of getting things through QA more quickly. I think we should do it this sprint though.