🐛 fix dashboard tile truncation logic (#237)
Fixes #236

Moved the truncation logic inside `DataCard` (mediated by props) because it handles the bolding internally. Alternatively we could make `DataCard` accept render props and move the logic outside, but this seems more complicated than necessary. The other alternative is to add more of the application logic inside the `DataCard` but this seems like a bad idea if we're trying to keep it as dumb as possible.

### Changes
- Add tests on `*StatsToProps` functions
- Modify `String.listify` to accept flag telling it not to insert `"and"` in the list -- defaults to the pre-existing behaviour
- Add another util to handle truncation
- `alternateBold` accepts an `offset` argument to avoid having to prepend an empty string to the argument sometimes
- Statistics functions now pass _all_ top values to the `DataCard` component, not just the first three, and the `DataCard` does the truncation.

### Testing Requirements
QA already passed

### Release Requirements
Can be released as soon as it is merged.

### Manual Deployment
- [ ] Dashboard App